Adding shadows

+++ RayTracerLib/Camera.cpp
@@ -21,8 +21,8 @@ namespace RayTracer {
   Ray Camera::ray_for_pixel(size_t x, size_t y) const
   {
     // Offset from the edge of the canvas to the pixel's center
-    auto xoffset = (x + 0.5) * _pixel_size;
-    auto yoffset = (y + 0.5) * _pixel_size;
+    auto xoffset = (x + 0.5f) * _pixel_size;
+    auto yoffset = (y + 0.5f) * _pixel_size;
 
     // Pixel coordinates in world space
     auto world_x = _half_width - xoffset;
@@ -44,7 +44,7 @@ namespace RayTracer {
 
     // MS version of OpenMP still requires the loop variable to be signed
     #pragma omp parallel for
-    for (long long x = 0; x < _hsize; ++x) {
+    for (long long x = 0; x < static_cast<long long>(_hsize); ++x) {
       for (size_t y = 0; y < _vsize; ++y) {
         auto r = ray_for_pixel(x, y);
         auto color = w.color_at(r);
